+++ vesta/clients.py
@@ -26,6 +26,7 @@ from typing import List
 from typing import Mapping
 from typing import Optional
 from typing import Union
+from typing import cast
 from urllib.parse import urljoin
 
 import requests
@@ -34,7 +35,7 @@ from .chars import COLS
 from .chars import ROWS
 from .chars import Rows
 
-__all__ = ["Client"]
+__all__ = ["Client", "LocalClient"]
 
 
 class Session(requests.Session):
@@ -58,7 +59,7 @@ class Client:
 
     Credentials must be provided as an ``api_key`` and ``api_secret``.
 
-    Optional, an alternate ``base_url`` can be specified, as well as any
+    Optionally, an alternate ``base_url`` can be specified, as well as any
     additional HTTP ``headers`` that should be sent with every request
     (such as a custom `User-Agent` header).
     """
@@ -101,7 +102,7 @@ class Client:
         subscription_id: str,
         message: Union[str, Rows],
     ) -> Dict[str, Any]:
-        """Post of a new message to a subscription.
+        """Post a new message to a subscription.
 
         The authenticated viewer must have access to the subscription.
 
@@ -134,3 +135,91 @@ class Client:
         )
         r.raise_for_status()
         return r.json()
+
+
+class LocalClient:
+    """Provides a Vestaboard Local API client interface.
+
+    A Local API key is required to read or write messages. This key is obtained
+    by enabling the Vestaboard's Local API using a Local API Enablement Token.
+
+    If you've already enabled your Vestaboard's Local API, that key can be
+    provided immediately. Otherwise, it can be set after the client is
+    constructed by calling :py:meth:`~enable` and assigning the resulting
+    Local API key to the :py:attr:`~api_key` property.
+
+    An alternate ``base_url`` can also be specified.
+
+    .. versionadded:: 0.8.0
+    """
+
+    def __init__(
+        self,
+        local_api_key: Optional[str] = None,
+        *,
+        base_url: str = "http://vestaboard.local:7000",
+    ):
+        self.session = Session(base_url)
+        if local_api_key is not None:
+            self.api_key = local_api_key
+
+    def __repr__(self):
+        return f"{type(self).__name__}(base_url={self.session.base_url!r})"
+
+    @property
+    def api_key(self) -> Optional[str]:
+        """The client's Local API key."""
+        return cast(
+            Optional[str],
+            self.session.headers.get("X-Vestaboard-Local-Api-Key"),
+        )
+
+    @api_key.setter
+    def api_key(self, value: str) -> None:
+        self.session.headers["X-Vestaboard-Local-Api-Key"] = value
+
+    @property
+    def enabled(self) -> bool:
+        """Check if a Local API key has been set, enabling the Local API."""
+        return self.api_key is not None
+
+    def enable(self, enablement_token) -> Dict[str, Any]:
+        """Enable the Vestaboard's Local API using a Local API Enablement Token.
+
+        If successful, the response will contain an API key that can be used to
+        read and write messages. Assign it to the :py:attr:`~api_key`
+        property before calling those methods.
+        """
+        r = self.session.post(
+            "/local-api/enablement",
+            headers={"X-Vestaboard-Local-Api-Enablement-Token": enablement_token},
+        )
+        r.raise_for_status()
+        return r.json()
+
+    def read_message(self) -> Optional[Rows]:
+        """Read the Vestaboard's current message."""
+        if not self.enabled:
+            raise RuntimeError("Local API has not been enabled")
+        r = self.session.get("/local-api/message")
+        r.raise_for_status()
+        try:
+            return r.json().get("message")
+        except requests.JSONDecodeError:
+            return None
+
+    def write_message(self, message: Rows) -> bool:
+        """Write a message to the Vestaboard.
+
+        `message` must be a two-dimensional (6, 22) array of character codes
+        representing the exact positions of characters on the board.
+
+        :raises ValueError: if `message` is a list with unsupported dimensions
+        """
+        if not self.enabled:
+            raise RuntimeError("Local API has not been enabled")
+        if len(message) != ROWS or not all(len(row) == COLS for row in message):
+            raise ValueError(f"expected a ({ROWS}, {COLS}) array of encoded characters")
+        r = self.session.post("/local-api/message", json=message)
+        r.raise_for_status()
+        return r.status_code == requests.codes.CREATED